Added HealthComponent to BP_PopPal and Simple GameOver UI
HealthComponent script is attachable to any actor. In this case it was attached to the character. Ball Enemies now apply damage to the character when they overlap colliders, which then decrements player health. If the player is hit 3 times, they "die" and a game over text is presented on the screen.