Updated FixedCamera Logic
Removed BP_FixedLocation and updated the camera spawning such that when a game starts, the camera will spawn at the same location of the actor that contains the tag "LevelFloor" in it. Future levels will need to be the same size to ensure the camera is the same every map.