Migrated from Deprecated Input System to Enhanced Input System; Added Camera Window Reference
Constraining aspect ratio and then pressing the fire input causes the letterboxes to shift and mess up the camera view. In an attempt to discover the underlying issue, switched from the deprecated input system to the enhanced input system. Also modeled a window reference in blender and imported into the project to use as a reference window for our camera view