Created BP_Window for Camera Setup
Created a BP_Window to use as a frame for our camera view. This window has a new static mesh imported from blender that has a fixed internal width of 32m and internal height of 16m, which is a 16:9 ratio. This is important because it allows us to setup our camera with the same view for every future level design and prevents users from modifying the camera view by resizing their window.