Setup Two Shot Power Up and Power Up Destruction
Created a IncreaseShotPowerUp script that subclasses BasePowerUp. It handles overlap collision with the player, so once the player comes into contact with it they are granted two projectiles to shoot for 6 seconds before they no longer have the power up. Whenever a power up is picked up the subclass of that power up will handle its destruction.