Added PowerUp Collision Detection; Added Double Jump Power Up
Player can now overlap the power-ups sphere collider to generate an overlap event and trigger the collided power up effect for a duration of 6 seconds. In addition to the power up that provides the player the ability to shoot two projectiles, I also added a power up that allows the player to jump twice.