Added BP_Shield PowerUp; Fixed Overlapping PowerUp Bug
Created Shield script that subclasses BasePowerUp, but power up effect yet to be implemented. Created temporary Blueprint for Shield script to create a placeable actor in the scene to test the power up. Fixed overlapping effects of picking up the same power up. Previously, when picking up the same power up in a small time span the effect would be cleared once the first duration was over, now the power up will persist until the last one's duration is finished.