Added Functionality to ShieldPowerUp
Now when a user picks up a shield power up, it will create a "bubble" shield around the player. That shield will remain in effect until the player is hit by an enemy, at which point the bubble will be destroyed and no damage will be taken by the player.