Create New Shield Material and Shield Break Particle Effect
Created a Vector Shield Material that has a panner animation to visualize a opaque shield protecting the player and applied to Bubble Shield Mesh. Created a new particle effect for the breaking of the shield when hit. Also adjusted colors of the external basic pickup asset pack to better match effects of power ups.