Updated Player Hit Notification (Flash)
Previously the visibility of the children static mesh components attached to the BP_RobotPal's character mesh wouldn't be toggled since the function call doesn't propagate to the children of the character mesh. Now, the visibility of all children static mesh's attached to the character mesh, EXCEPT for the shield will toggle as expected.