Added a LootSystem Class
When the RoboQuest game mode is initialized, it will create a single instance of this object of class LootSystem. This class contains a pre-declared array of droppable loot, all of type FItem (struct). Each FItem contains two properties, one is the subclass of the powerup to drop and the second is the drop rate of that item. Lootsystem also contains a function for calculating a random drop rate and based on its comparison to each items drop rate chance, it will return the subclass of that item to spawn in its functions out parameter. Will update this in the next commit to spawn the intended loot.